+++ Sources/Tests/CreationTestCase.swift
@@ -17,26 +17,26 @@ class FooCell : UICollectionViewCell, ModelTransfer
     }
 }
 
-class NonOptionalCollectionViewController : UIViewController, DTCollectionViewNonOptionalManageable {
-    var collectionView: UICollectionView!
+class OptionalCollectionViewController : UIViewController, DTCollectionViewManageable {
+    var optionalCollectionView: UICollectionView?
 }
 
 class CreationTestCase: XCTestCase {
     
-    func testManagingWithNonOptionalCollectionViewWorks() {
-        let controller = NonOptionalCollectionViewController()
-        controller.collectionView = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewLayout())
+    func testManagingWithOptionalCollectionViewWorks() {
+        let controller = OptionalCollectionViewController()
+        controller.optionalCollectionView = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewLayout())
         XCTAssertTrue(controller.manager.isManagingCollectionView)
     }
     
     func testDelegateIsNotNil() {
         let controller = DTCellTestCollectionController()
-        XCTAssertNotNil(controller.manager.storage.delegate)
+        XCTAssertNotNil((controller.manager.storage as? BaseUpdateDeliveringStorage)?.delegate)
     }
     
     func testDelegateIsNotNilForMemoryStorage() {
         let controller = DTCellTestCollectionController()
-        XCTAssertNotNil(controller.manager.storage.delegate)
+        XCTAssertNotNil((controller.manager.storage as? BaseUpdateDeliveringStorage)?.delegate)
     }
     
     func testSwitchingStorages() {
@@ -61,7 +61,6 @@ class CreationTestCase: XCTestCase {
     func testCreatingCollectionControllerFromXIB()
     {
         let controller = XibCollectionViewController(nibName: "XibCollectionViewController", bundle: Bundle(for: type(of: self)))
-        let _ = controller.view
         controller.manager.register(FooCell.self)
     }